+++ src/js/main.js
@@ -1,4 +1,6 @@
 import '../scss/style.scss';
+import clothes from './clothes';
+import accessories from './accessories';
 
 // search-input
 
@@ -11,120 +13,6 @@ btn.addEventListener('click', () => {
 });
 
 // cards
-
-const clothes = [
-    {
-        id:    0,
-        img:   'src/assets/image04.jpg',
-        alt:   'Футболка "Эволюционируй или сдохни"',
-        title: 'Футболка "Эволюционируй или сдохни"',
-        price: 220,
-        badge: true,
-        size:  'S/M/L',
-    },
-    {
-        id:    1,
-        img:   'src/assets/image11.jpg',
-        alt:   'Футболка "Все очень круто, но надо переделать"',
-        title: 'Футболка "Все очень круто, но надо переделать"',
-        price: 200,
-        badge: true,
-        size:  'S/M/L',
-    },
-    {
-        id:    5,
-        img:   'src/assets/image12.jpg',
-        alt:   'Толстовка "Kolesa Team"',
-        title: 'Толстовка "Kolesa Team"',
-        price: 500,
-        badge: false,
-        size:  'S/M/L',
-    },
-    {
-        id:    6,
-        img:   'src/assets/image13.jpg',
-        alt:   'Толстовка "Kolesa Team"',
-        title: 'Толстовка "Kolesa Team"',
-        price: 510,
-        badge: false,
-        size:  'S/M/L',
-    },
-    {
-        id:    7,
-        img:   'src/assets/image14.jpg',
-        alt:   'Толстовка "Kolesa Team"',
-        title: 'Толстовка "Kolesa Team"',
-        price: 520,
-        badge: false,
-        size:  'S/M/L',
-    },
-    {
-        id:    8,
-        img:   'src/assets/image17.png',
-        alt:   'Толстовка "Kolesa Team"',
-        title: 'Толстовка "Kolesa Team"',
-        price: 525,
-        badge: false,
-        size:  'S/M/L',
-    },
-];
-
-const accessories = [
-    {
-        id:    11,
-        img:   'src/assets/image07.jpg',
-        alt:   'Бутылка для воды',
-        title: 'Бутылка для воды',
-        price: 150,
-        badge: false,
-        size:  'M',
-    },
-    {
-        id:    13,
-        img:   'src/assets/image15.jpg',
-        alt:   'Рюкзак "Kolesa Team"',
-        title: 'Рюкзак "Kolesa Team"',
-        price: 400,
-        badge: true,
-        size:  'S/M/L',
-    },
-    {
-        id:    14,
-        img:   'src/assets/image16.jpg',
-        alt:   'Рюкзак "Kolesa Team"',
-        title: 'Рюкзак "Kolesa Team"',
-        price: 390,
-        badge: true,
-        size:  'S/M/L',
-    },
-    {
-        id:    15,
-        img:   'src/assets/image08.jpg',
-        alt:   'Термо-кружка "Kolesa Team"',
-        title: 'Термо-кружка "Kolesa Team"',
-        price: 380,
-        badge: false,
-        size:  'S/M/L',
-    },
-    {
-        id:    16,
-        img:   'src/assets/image09.jpg',
-        alt:   'Вязанная шапка "Kolesa Team"',
-        title: 'Вязанная шапка "Kolesa Team"',
-        price: 230,
-        badge: false,
-        size:  'S/M/L',
-    },
-    {
-        id:    17,
-        img:   'src/assets/image20.jpg',
-        alt:   'Солнцезащитные очки',
-        title: 'Солнцезащитные очки',
-        price: 100,
-        badge: false,
-        size:  'S/M/L',
-    },
-];
 
 const all = clothes.concat(accessories);
 
@@ -144,8 +32,8 @@ const makeProductCard = (img, alt, title, price, badge, size, id) => ` <div data
     </div>`;
 
 const renderCard = (item) => {
-    const wrapper = document.createElement('div');
     const filterItem = item.sort(good => (good.badge ? -1 : 1));
+    const wrapper = document.createElement('div');
 
     wrapper.className = 'content__products';
     filterItem.forEach((card) => {
@@ -197,7 +85,7 @@ function openModal(id) {
     document.querySelector('.modal-main-wrap').classList.add('modal-main-wrap--active');
 }
 
-document.querySelectorAll('[data-id]').forEach((item) => {
+document.querySelectorAll('div[data-id]').forEach((item) => {
     item.addEventListener('click', () => {
         openModal(item.dataset.id);
     });
